XW-1510 | upgrade ember-cli-mirage and use model + serializer
