+++ front/main/mirage/fixtures/curated-content-editor-item.js
@@ -1,0 +1,107 @@
+/* eslint-disable max-len */
+
+export default [
+	{
+		label: 'Section 1',
+		image_id: 2202,
+		items: [
+			{
+				article_id: 2216,
+				image_id: 2216,
+				items: '',
+				label: 'Category Test',
+				title: 'Category:Test',
+				type: 'category',
+				image_url: 'http://vignette2.wikia.nocookie.net/concf/images/1/13/DSC_0211.JPG/revision/latest/window-crop/width/50/x-offset/841/y-offset/0/window-width/2161/window-height/2160?cb=20150730090340',
+				node_type: 'item'
+			}
+		],
+		image_crop: {
+			square: {
+				x: 92,
+				y: 0,
+				width: 412,
+				height: 412
+			}
+		},
+		image_url: 'http://vignette2.wikia.nocookie.net/concf/images/7/76/Satisfied-Seal.jpg/revision/latest?cb=20150626171201',
+		node_type: 'section'
+	},
+	{
+		label: 'Section 2',
+		image_id: 2361,
+		items: [
+			{
+				article_id: 2303,
+				image_id: 2303,
+				items: '',
+				label: 'Category Test 2',
+				title: 'Category:Test2',
+				type: 'category',
+				image_url: 'http://vignette2.wikia.nocookie.net/concf/images/e/e4/070521_hamster_02.jpg/revision/latest/window-crop/width/50/x-offset/56/y-offset/0/window-width/466/window-height/465?cb=20150408114003',
+				node_type: 'item'
+			}
+		],
+		image_crop: {
+			square: {
+				x: 114,
+				y: 0,
+				width: 381,
+				height: 381
+			}
+		},
+		image_url: 'http://vignette2.wikia.nocookie.net/concf/images/1/1e/Leadership-council-icon-oxfam-america-1_610x381.png/revision/latest?cb=20160422112542',
+		node_type: 'section'
+	},
+	{
+		label: 'Featured Content',
+		image_id: 0,
+		items: [
+			{
+				article_id: 2178,
+				image_id: 2364,
+				image_crop: {
+					landscape: {
+						x: 0,
+						y: 699,
+						width: 2344,
+						height: 1319
+					}
+				},
+				items: '',
+				label: 'Featured item',
+				title: 'Test',
+				type: 'article',
+				image_url: 'http://vignette4.wikia.nocookie.net/concf/images/b/b1/Freddie_mercury.jpg/revision/latest?cb=20160505132527',
+				node_type: 'item'
+			}
+		],
+		featured: 'true',
+		node_type: 'section'
+	},
+	{
+		label: '',
+		image_id: 0,
+		items: [
+			{
+				article_id: 1766,
+				image_id: 2182,
+				image_crop: {
+					square: {
+						x: 0,
+						y: 0,
+						width: 620,
+						height: 620
+					}
+				},
+				items: '',
+				label: 'Optional category',
+				title: 'Category:Test3',
+				type: 'category',
+				image_url: 'http://vignette1.wikia.nocookie.net/concf/images/9/97/Doge.jpg/revision/latest?cb=20150622134753',
+				node_type: 'item'
+			}
+		],
+		node_type: 'section'
+	}
+];
