+++ front/main/mirage/models/curated-content-editor-item.js
@@ -1,0 +1,3 @@
+import {Model} from 'ember-cli-mirage';
+
+export default Model;
