+++ front/main/mirage/serializers/application.js
@@ -1,0 +1,3 @@
+import {Serializer} from 'ember-cli-mirage';
+
+export default Serializer.extend();
