+++ front/main/package.json
@@ -38,7 +38,7 @@
     "ember-cli-htmlbars-inline-precompile": "0.3.1",
     "ember-cli-inject-live-reload": "1.4.0",
     "ember-cli-inline-content": "0.4.0",
-    "ember-cli-mirage": "0.1.13",
+    "ember-cli-mirage": "0.2.0-beta.9",
     "ember-cli-moment-shim": "1.0.0",
     "ember-cli-qunit": "1.4.0",
     "ember-cli-sass": "5.2.1",
